backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.6.0
imagehash==4.3.1
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
python-dotenv==1.0.0
werkzeug==3.0.1
